backend/requirements.txt
flask
flask-cors
scikit-learn
numpy
pandas
google-generativeai
gunicorn

